+++ .gitignore
@@ -1,6 +1,8 @@
-/app/torrent
+/torrent
 /config/config.php
 /php-7.2.12-nts
 /*.bat
 /*.sh
 /vendor
+.idea
+.php_cs.cache
