+++ app/Client/AbstractClientInterface.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: Rhilip
+ * Date: 1/17/2020
+ * Time: 2020
+ */
+
+namespace IYUU\Client;
+
+interface AbstractClientInterface
+{
+}
