--- config/version.php
@@ -1,1 +1,0 @@
-<?php return '20191224.1010';
