+++ init.php
@@ -3,9 +3,11 @@
 // 公共入口文件
 //----------------------------------
 // 定义目录
+use IYUU\AutoReseed;
+
 defined('ROOT_PATH') or define("ROOT_PATH", __DIR__);
 define('DS', DIRECTORY_SEPARATOR);
-define('TORRENT_PATH', APP_PATH.DS.'torrent'.DS);
+define('TORRENT_PATH', ROOT_PATH.DS.'torrent'.DS);
 
 // 严格开发模式
 error_reporting(E_ALL);
@@ -14,6 +16,7 @@ error_reporting(E_ALL);
 // 永不超时
 ini_set('max_execution_time', 0);
 set_time_limit(0);
+
 // 内存限制，如果外面设置的内存比 /etc/php/php-cli.ini 大，就不要设置了
 if (intval(ini_get("memory_limit")) < 1024) {
     ini_set('memory_limit', '1024M');
@@ -36,3 +39,12 @@ if (file_exists(ROOT_PATH."/config/config.php")) {
 }
 
 require_once ROOT_PATH . '/vendor/autoload.php';
+
+AutoReseed::init();
+$hashArray = AutoReseed::get();
+if (AutoReseed::$move != null) {
+    echo "种子移动完毕，请重新编辑配置，再尝试辅种！ \n\n";
+    exit;
+}
+AutoReseed::call($hashArray);
+AutoReseed::wechatMessage();
